+++ workers/worker-python/src/worker_python/engines/__init__.py
@@ -1,4 +1,4 @@
-from openregex_libs.models import WorkerInfo
+from openregex_libs.models import WorkerInfo, WORKER_SCHEMA_VERSION
 from .common import WORKER_VERSION, WORKER_RELEASE_DATE
 from .python_re import engine as python_re
 from .python_regex import engine as python_regex
@@ -12,5 +12,6 @@ WORKER_INFO = WorkerInfo(
     worker_name="worker-python",
     worker_version=WORKER_VERSION,
     worker_release_date=WORKER_RELEASE_DATE,
+    worker_schema_version=WORKER_SCHEMA_VERSION,
     engines=ENGINES
 )